collectMetadata {DbDiagnostics}	R Documentation

collectMetadata

Description

collectMetadata

Usage

collectMetadata(
  outputFolder,
  outputFileName = "metadata",
  siteName = NA,
  siteOHDSIParticipation = NA,
  siteOHDSIRunPackage = NA,
  dataFullName = NA,
  dataShortName = NA,
  dataContactName = NA,
  dataContactEmail = NA,
  dataDoiType = NA,
  governanceTime = NA,
  dataProvenance = NA,
  refreshTime = NA
)

Arguments

outputFolder
The output folder location where the csv with the metadata should be saved

outputFileName
The name of the file to be written. Default is metadata.csv

siteName
The name of the site or institution that owns or licenses the data.

siteOHDSIParticipation
Yes/No if the site contributed to an OHDSI study in the past

siteOHDSIRunPackage
Yes/No if site has someone who can run and/or debug an OHDSI study package

dataFullName
The full name of the database

dataShortName
The short name or nickname of the database

dataContactName
The name of the person who should be contacted in the event this database is identified as a good candidate for a study

dataContactEmail
The email address of the person who should be contacted in the event this database is identified as a good candidate for a study

dataDoiType
The type of data object identifier (DOI) the database has. Options are "DOI", "CURIE", "ARK", "Other",

governanceTime
How long (in weeks) it typically takes to receive approval to run a study on this database

dataProvenance
The type(s) of data that are present in your database. Options are "Electronic Health Records", "Administrative Claims", "Disease-specific Registry", "Wearable or Sensor Data", "Other".

refreshTime
How often the data are refreshed

[Package DbDiagnostics version 1.3.1 Index]