executeDbDiagnostics {DbDiagnostics}	R Documentation

executeDbDiagnostics

Description

executeDbDiagnostics

Usage

executeDbDiagnostics(
  connectionDetails,
  resultsDatabaseSchema,
  resultsTableName,
  outputFolder = getwd(),
  dataDiagnosticsSettingsList
)

Arguments

connectionDetails
A connectionDetails object for connecting to the database containing the DbProfile results

resultsDatabaseSchema
The fully qualified database name of the results schema where the DbProfile results are housed. Default is "dp_temp".

resultsTableName
The name of the table in the results schema with the DbProfile results. Default is "dp_achilles_results_augmented."

outputFolder
Results will be written to this directory, default = getwd()

dataDiagnosticsSettingsList
A list of settings objects, each created from DataDiagnostics::createDataDiagnosticsSettings() function and each representing one analysis.

[Package DbDiagnostics version 1.3.1 Index]