createDataDiagnosticsSettings {DbDiagnostics}	R Documentation

createDataDiagnosticsSettings

Description

createDataDiagnosticsSettings

Usage

createDataDiagnosticsSettings(
  analysisId,
  analysisName = "analysis",
  minAge = NULL,
  maxAge = NULL,
  genderConceptIds = c(8507, 8532),
  raceConceptIds = NULL,
  ethnicityConceptIds = NULL,
  studyStartDate = NULL,
  studyEndDate = NULL,
  requiredDurationDays = 365,
  requiredDomains = c("condition", "drug"),
  desiredDomains = NULL,
  requiredVisits = NULL,
  desiredVisits = NULL,
  targetName = "target",
  targetConceptIds = NULL,
  comparatorName = NULL,
  comparatorConceptIds = NULL,
  indicationName = NULL,
  indicationConceptIds = NULL,
  includeIndicationInCalc = F,
  outcomeName = NULL,
  outcomeConceptIds = NULL
)

Arguments

analysisId
The identifier for the specific analysis.

analysisName
The name for the specific analysis.

minAge
The minimum age for patients included in the analysis. Default is the lowest age available in the database.

maxAge
The minimum age for patients included in the analysis. Default is the highest age available in the database.

genderConceptIds
The required sex(es) at birth for patients included in the analysis, expressed as a vector of concepts. List the standard OMOP concepts for the required values as found here (https://athena.ohdsi.org/search-terms/terms?domain=Gender&standardConcept=Standard&page=1&pageSize=15&query=). Default is c(8507, 8532) for male and female.

raceConceptIds
The required races for patients included in the analysis, expressed as a vector of concepts, e.g. c(8515,8527). If no restriction leave blank, else list the standard OMOP concepts for the required values as found here (https://athena.ohdsi.org/search-terms/terms?domain=Race&standardConcept=Standard&page=1&pageSize=15&query=).

ethnicityConceptIds
The required ethnicities for patients included in the analysis, expressed as a vector of concepts, e.g. c(38003564,38003563) If no restriction leave blank, else list the standard OMOP concepts for the required values as found here (https://athena.ohdsi.org/search-terms/terms?domain=Ethnicity&standardConcept=Standard&page=1&pageSize=15&query=).

studyStartDate
The start date of the analysis. Date format it 'YYYYMM'. If no restriction leave blank.

studyEndDate
The end date of the analysis. Date format it 'YYYYMM'. If no restriction leave blank.

requiredDurationDays
The minimum required follow-up time in days for patients included in the analysis.

requiredDomains
The data domains required for ALL patients in the analysis, expressed as a character vector. Valid values are condition,drug,device,measurement,measurementValues,death,procedure,observation. Default is c("condition", "drug")

desiredDomains
The domains required for SOME patients in the analysis. Meaning you need the dataset to have these data but each person is not required to have a record. This is most often used to identify the presence of outcomes of interest. Valid values are condition,drug,device,measurement,measurementValues,death,procedure,observation. If no restriction leave blank.

requiredVisits
The visits required for ALL patients in the study, expressed as a character vector, e.g. c("IP","OP"). Valid values are IP,OP,ER. If no restriction leave blank.

desiredVisits
The visits required for SOME patients in the study, expressed as a character vector. Meaning you need the dataset to have these data but each person is not required to have a record. This is most often used to identify the presence of outcomes of interest. Valid values are IP,OP,ER.

targetName
The name of the target of interest.

targetConceptIds
A vector containing the required target concepts.

comparatorName
The name of the comparator of interest.

comparatorConceptIds
A vector containing the required comparator concepts.

indicationName
The name of the indication of interest.

indicationConceptIds
A vector containing the required indication concepts.

includeIndicationInCalc
A T/F object indicating whether the proportion of people with the indication concepts should be included in the estimated sample size calculation. Default is FALSE.

outcomeName
The name of the outcome of interest.

outcomeConceptIds
A vector containing the required outcome concepts.

[Package DbDiagnostics version 1.3.1 Index]